+++ CHANGELOG.md
@@ -20,6 +20,8 @@ Template:
 ## [1.0.2] - Unreleased - 2023-10-17
 
 ### Changed
+- distributed strategy: charge vehicles with any strategy at depot and opportunity stations
+- distributed strategy: battery behavior adapted
 - new strategies: peak_load_window
 - plotting time windows with shaded backgrounds
 - changed total power plot: remove individual loads, add sum of batteries, add GC load